requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
paho-mqtt
influxdb-client
requests
python-dotenv
psycopg2-binary